# docs/VMOGD3swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "VMOGD3Project",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/api/v1",
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "description": "",
      "name": "token",
      "in": "header"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/admins/{id}": {
      "get": {
        "summary": "Get admin by id",
        "tags": [
          "Admin"
        ],
        "operationId": "Getadminbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Update admin",
        "tags": [
          "Admin"
        ],
        "operationId": "Updateadmin",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateadminRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete admin",
        "tags": [
          "Admin"
        ],
        "operationId": "Deleteadmin",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DeleteadminRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/employees/{id}": {
      "get": {
        "summary": "Get employee by id",
        "tags": [
          "Employee"
        ],
        "operationId": "Getemployeebyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Update employee",
        "tags": [
          "Employee"
        ],
        "operationId": "Updateemployee",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateemployeeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "patch": {
        "summary": "Set admin for employee",
        "tags": [
          "Employee"
        ],
        "operationId": "Setadminforemployee",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SetadminforemployeeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete employee",
        "tags": [
          "Employee"
        ],
        "operationId": "Deleteemployee",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateadminRequest": {
      "title": "UpdateadminRequest",
      "example": {
        "username": "empladmin",
        "password": "12345678"
      },
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "username",
        "password"
      ]
    },
    "DeleteadminRequest": {
      "title": "DeleteadminRequest",
      "example": {
        "status": 1
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "status"
      ]
    },
...
    "UpdateemployeeRequest": {
      "title": "UpdateemployeeRequest",
      "example": {
        "name": "Employee fix",
        "dob": "10/20/1994",
        "address": "19 Duy Tan",
        "idNumber": "0590",
        "phone": "0901523596",
        "experienceYear": 5,
        "foreignLanguage": "English, Japanese",
        "certificate": "Senior Dev",
        "teches": [
          "d9172cad-0ba6-48d8-8a2b-824853a54035",
          "6ca35f3f-322c-467b-8349-efac19b9332c"
        ]
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "dob": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "idNumber": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "experienceYear": {
          "type": "integer",
          "format": "int32"
        },
        "foreignLanguage": {
          "type": "string"
        },
        "certificate": {
          "type": "string"
        },
        "teches": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "name",
        "dob",
        "address",
        "idNumber",
        "phone",
        "experienceYear",
        "foreignLanguage",
        "certificate",
        "teches"
      ]
    },
    "SetadminforemployeeRequest": {
      "title": "SetadminforemployeeRequest",
      "example": {
        "isAdmin": 0
      },
      "type": "object",
      "properties": {
        "isAdmin": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "isAdmin"
      ]
    },
...
  },
  "security": [
    {
      "token": []
    }
  ],
...
}